# backend/openapi.yaml
openapi: '3.0.0'
info:
  title: Sorikama API Gateway - Hub Central
  version: '1.0.0'
  description: API Gateway centralisée pour l'écosystème Sorikama
  contact:
    name: Support Sorikama
    url: https://sorikama.com/support
    email: support@sorikama.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:7000/api/v1
    description: Serveur de développement local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT d'authentification
...
paths:
...
  /service-user/profile:
    get:
      tags: [Services Externes]
      summary: Obtenir le profil utilisateur (pour services externes)
      security: []
      parameters:
        - in: header
          name: X-Service-Token
          required: true
          schema:
            type: string
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Profil utilisateur
    patch:
      tags: [Services Externes]
      summary: Mettre à jour le profil utilisateur (pour services externes)
      security: []
      parameters:
        - in: header
          name: X-Service-Token
          required: true
          schema:
            type: string
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName: { type: string }
                lastName: { type: string }
      responses:
        '200':
          description: Profil mis à jour
  /service-user/password:
    put:
      tags: [Services Externes]
      summary: Changer le mot de passe (pour services externes)
      security: []
      parameters:
        - in: header
          name: X-Service-Token
          required: true
          schema:
            type: string
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword: { type: string }
      responses:
        '200':
          description: Mot de passe changé

  # === SERVICE CALLBACK (pour les services externes) ===
...
security:
  - bearerAuth: []